Installed app on my wifes phone and watch her use the user controls. Made notes of areas for improvement.
This morning I changed the textboxes to use Google material so that hint text floats when user is typing and they don't forget the form box name. Also made use of the textbox error feature which highlights the text.

Fixed the email validation as well.

Added check that username has no spaces